Modifying Frames2MatrixConverter parameters and generalizing main.py
For frames2matrix: Changed the clear_frame input into clear_frame_number and added a parameter for the number of frames just to make the code look nicer for me.

For main.py: I added variables that tell where all the inputs, intermediates, and outputs of the program will go, allowing for them to also be changed. Previously there was a purely static folder structure but now it can be varied without having to modify the code too much by hand, just by changing one parameter.