Add two types of converters
The full_converter is a function that takes in all parameters that would be needed in the program at once in the function call and can be used if the user already knows all the values they need to convert.  If is it missing a value that it needs, it will throw an error.

The partial_converter can take in as many or as few parameters in the function call as the user gives and will prompt the user to give all the values that they did not know/give when calling the function. It also has some default parameters that the user can use. It will show frames in the user needs to input any information that needs to be gotten from the frames.

Updated main.py so it just calls the converters

Since all of what main.py did has been moved to the converters, you can just have main.py call the converters.