Changed the way to make the column2note dictionary
Now, it takes in the column of the first and tenth white notes and from them extrapolates all the other notes on the piano and their column.

This makes it more consistent as it can be done even without a clear frame or without consistent coloring on the keys on the frame it would choose to read, as it does not even look at any frame itself.

One downside is this assumes that the keys are evenly spaced, however this is nearly always true.